Create the grid of a virtual environment

Create the grid of a virtual environment

Usage

create_grid(
  ymin = 0,
  ymax = 90,
  xmin = 0,
  xmax = 90,
  step_x = 0.5,
  step_y = 0.5
)

Arguments

ymin

Minimum latitude (default to 0)

ymax

Maximum latitude (default to 90)

xmin

Minimum longitude (default to 0)

xmax

Maximum latitude (default to 90)

step_x

Longitude steps to be used to generate points (default to 0.5)

step_y

Latitude steps to be used to generate points (default to 0.5)

Value

A grid in dataframe format.

Examples

grid <- create_grid()